+++ tt_metal/impl/dispatch/device_command_calculator.hpp
@@ -73,8 +73,10 @@ public:
 
     // Calculator sizing for CQ_DISPATCH_CMD_WRITE_LINEAR_H (dispatch_h linear write)
     // Mirrors add_dispatch_write_linear for sizing/alignment purposes.
+    // Argument data_sizeB can be uint32_t, as it is relevant only when data will be inline (since it will be limited to
+    // size of prefetch cmddat_q). However defining it uint64_t type to avoid confusion.
     template <bool flush_prefetch = true, bool inline_data = false>
-    void add_dispatch_write_linear_h(uint32_t data_sizeB) {
+    void add_dispatch_write_linear_h(uint64_t data_sizeB) {
         this->add_prefetch_relay_inline();
         this->cmd_write_offsetB += sizeof(CQDispatchCmdLarge);
